+++ tests/TestCase/View/Helper/FormHelperTest.php
@@ -64,6 +64,8 @@ class FormHelperTest extends HelperTestCase
 
         $expected = ['button' => ['type' => 'button', 'class' => 'btn btn-danger'], $title, '/button'];
         $this->assertHtml($expected, $this->Helper->button($title, ['class' => 'btn-danger']));
+
+        $this->assertSame('<button class="btn btn-primary" type="button"></button>', $this->Helper->button());
     }
 
     /**
@@ -534,6 +536,15 @@ class FormHelperTest extends HelperTestCase
             '/a',
         ];
         $this->assertHtml($expected, $this->Helper->postButton($title, $url, ['class' => 'btn-danger']));
+
+        $expected = [
+            'form' => ['name', 'style' => 'display:none;', 'method' => 'post', 'action' => '/'],
+            'input' => ['type' => 'hidden', 'name' => '_method', 'value' => 'POST'],
+            '/form',
+            'a' => ['href' => '#', 'class' => 'btn btn-light', 'onclick', 'role' => 'button', 'title' => ''],
+            '/a',
+        ];
+        $this->assertHtml($expected, $this->Helper->postButton());
     }
 
     /**
@@ -578,6 +589,15 @@ class FormHelperTest extends HelperTestCase
             '/a',
         ];
         $this->assertHtml($expected, $this->Helper->postLink('" \'', $url));
+
+        $expected = [
+            'form' => ['name', 'style' => 'display:none;', 'method' => 'post', 'action' => '/'],
+            'input' => ['type' => 'hidden', 'name' => '_method', 'value' => 'POST'],
+            '/form',
+            'a' => ['href' => '#', 'title' => '', 'onclick'],
+            '/a',
+        ];
+        $this->assertHtml($expected, $this->Helper->postLink());
     }
 
     /**